# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hellfire Gatekeeper API — Listados (productos y órdenes)
  version: 1.0.0
  license:
    name: Proprietary
    url: https://github.com/radamesvaz/bakery-app
  description: |
    Contrato de **paginación por cursor** y **búsqueda/filtros** para listados públicos de catálogo y listado autenticado de pedidos.

    ## Respuesta de listado (envelope)

    Todos los endpoints de listado devuelven un **objeto JSON** (no un array en la raíz):

    ```json
    { "items": [ ... ], "next_cursor": "<string | null>" }
    ```

    - `items`: filas de la página actual.
    - `next_cursor`: si es **no nulo**, hay más resultados; en la siguiente petición envía el mismo valor en el query param `cursor`. Si es **null**, no hay página siguiente.

    ## Regla al cambiar filtros de búsqueda

    Si cambias **`q`** (productos) o **`id_user`** / **`status`** / **`ignore_status`** (órdenes), **no reutilices** un `cursor` obtenido con otra combinación de parámetros. Vuelve a llamar **sin** `cursor` (primera página del nuevo filtro).

    ## Formatos de cursor (distintos por recurso)

    **Productos:** cursor opaco **versión 1**, keyset sobre `id_product` descendente. Solo aplica al listado de productos.

    **Órdenes:** cursor opaco **versión 2**, keyset sobre `(created_on ASC, id_order ASC)`. **No** es intercambiable con el de productos.

...

    ## Errores HTTP 400

    El servidor suele responder con cuerpo **text/plain** (mensaje breve), no JSON. Casos frecuentes:

...
servers:
...
paths:
...
  /auth/products/{id}:
    get:
      tags: [Catalog]
      summary: Obtener producto por ID (admin, cualquier estado)
      description: |
        Devuelve el producto aunque esté `inactive` o `deleted`.
        Requiere **Bearer JWT** con rol admin o superadmin.
      operationId: getProductByIdAdmin
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Producto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: JWT ausente o inválido
        "403":
          description: Rol insuficiente (no admin/superadmin)
        "404":
          description: Producto no encontrado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Product:
      type: object
      properties:
        id_product:
          type: integer
          format: int64
        tenant_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        track_inventory:
          type: boolean
          description: |
            When true, stock is decremented on order create and reverted on admin cancel.
            When false, stock is not enforced (unlimited / service-style products).
        stock:
          type: integer
          format: int64
        status:
          type: string
          description: |
            Product lifecycle. Public catalog endpoints only return `active`.
            `inactive` hides from catalog but keeps the product; `deleted` soft-deletes
            (admin may also clean up images).
          enum: [active, inactive, deleted]
        image_urls:
          type: array
          items:
            type: string
        thumbnail_url:
          type: string
        created_on:
          type: string
          format: date-time
          nullable: true
...